+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. Dates are d
 
 #### [Unreleased](https://github.com/hougesen/mdsf/compare/v0.8.5...HEAD)
 
+- test: format --config path argument [`#942`](https://github.com/hougesen/mdsf/pull/942)
 - refactor(cli): make --log-level a global argument [`#941`](https://github.com/hougesen/mdsf/pull/941)
 - feat(init): add --force argument [`#940`](https://github.com/hougesen/mdsf/pull/940)
 - test: integration test of cli [`#938`](https://github.com/hougesen/mdsf/pull/938)